requirements.txt
Django==4.2.7
djangorestframework==3.14.0
pandas>=2.2.0
openpyxl==3.1.2
python-decouple==3.8
whitenoise==6.6.0
gunicorn==21.2.0
psycopg2-binary>=2.9.0
dj-database-url>=2.1.0

